Add authorization

+++ AdvertisementPortal/AdvertisementPortal/Services/AdvertisementService.cs
@@ -1,10 +1,13 @@
-﻿using AdvertisementPortal.Entities;
+﻿using AdvertisementPortal.Authorization;
+using AdvertisementPortal.Entities;
 using AdvertisementPortal.Exceptions;
 using AdvertisementPortal.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace AdvertisementPortal.Services
 {
@@ -13,12 +16,15 @@ namespace AdvertisementPortal.Services
         private readonly AdvertisementDbContext dbContext;
         private readonly IMapper mapper;
         private readonly ILogger<AdvertisementService> logger;
+        private readonly IAuthorizationService authorizationService;
 
-        public AdvertisementService(AdvertisementDbContext dbContext, IMapper mapper, ILogger<AdvertisementService> logger)
+        public AdvertisementService(AdvertisementDbContext dbContext, IMapper mapper, ILogger<AdvertisementService> logger,
+            IAuthorizationService authorizationService)
         {
             this.dbContext = dbContext;
             this.mapper = mapper;
             this.logger = logger;
+            this.authorizationService = authorizationService;
         }
 
         public AdvertisementDto GetById(int id)
@@ -41,16 +47,17 @@ namespace AdvertisementPortal.Services
             return advertisementDto;
         }
 
-        public int Create(CreateAdvertisementDto createDto)
+        public int Create(CreateAdvertisementDto createDto, int userId)
         {
             var advertisement = mapper.Map<Advertisement>(createDto);
+            advertisement.CreatedById = userId;
             dbContext.Advertisements.Add(advertisement);
             dbContext.SaveChanges();
 
             return advertisement.Id;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             logger.LogError($"Advertisement with id: {id} delete");
 
@@ -58,15 +65,31 @@ namespace AdvertisementPortal.Services
 
             if (advertisement is null) throw new NotFoundException("Advertisement not found");
 
+            var authorizationResult = authorizationService.AuthorizeAsync(user, advertisement,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
             dbContext.Advertisements.Remove(advertisement);
             dbContext.SaveChanges();
         }
 
-        public void Update(int id, UpdateAdvertisementDto updateDto)
+        public void Update(int id, UpdateAdvertisementDto updateDto, ClaimsPrincipal user)
         {
             var advertisement = dbContext.Advertisements.FirstOrDefault(a => a.Id == id);
 
             if (advertisement is null) throw new NotFoundException("Advertisement not found");
+
+            var authorizationResult = authorizationService.AuthorizeAsync(user, advertisement,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             advertisement.Title = updateDto.Title;
             advertisement.Content = updateDto.Content;
